[day07] part 2, ugh
I guess a better way would be to properly implement a tree (trie) or
just use a module for it. The way I have implemented it here is that I
only have a dictionary which tells me for each node what is the nodes
weight and who are the nodes children. Nothing is nested, so going through
the collection, finding neighbors (same-level nodes) is cumbersome.

Also, readability sucks. `dict[node][0]` is the node weight, `dict[node][1]`
is the list of node children. So the value of every key in the dict is a
list which contains an integer and a list of strings.

Anyway, I am looking forward to rewriting this and learning something.