[day16] nailed the repeating dance patterns!
no need to run the simulation for one billion dances. I figured this out
by running one thousand dances and putting the resulting strings in a
set. the length of the resulting set was 42, which means that there are
most probably only 42 unique dances in the simulation which repeat
periodically. I tried that and it worked!